Fixed Bugs app working e2e

+++ ecommerce/src/main/java/Hs/Ecommerce/Core/DTO/Request/RatingDTO.java
@@ -3,7 +3,7 @@ package Hs.Ecommerce.Core.DTO.Request;
 public record RatingDTO(
         Long userId,
         Long productId,
-        Long rating,
+        Double rating,
         String review
 ) {
 }
